% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classify-time-series-data.R
\name{classify_time_series_data}
\alias{classify_time_series_data}
\title{Classify Time Series Data}
\usage{
classify_time_series_data(
  data,
  ...,
  date_time = "date_time",
  value = "value",
  questionable_min = 0,
  questionable_max = 30,
  erroneous_min = -0.5,
  erroneous_max = 40,
  questionable_rate = 2,
  erroneous_rate = 5,
  questionable_buffer = 1,
  erroneous_buffer = 1,
  gap_range = 5
)
}
\arguments{
\item{data}{A data frame.}

\item{questionable_min}{A numeric value indicating the lower bound of the
questionable range of temperature values.}

\item{questionable_max}{A numeric value indicating the upper bound of the
questionable range of temperature values.}

\item{erroneous_min}{A numeric value indicating the lower bound of the
erroneous range of temperature values.}

\item{erroneous_max}{A numeric value indicating the upper bound of the
erroneous range of temperature values.}

\item{questionable_rate}{A numeric value indicating the rate of change
(temperature per hour) of temperature values that is considered
questionable.}

\item{erroneous_rate}{A numeric value indicating the rate of change
(temperature per hour) of temperature values that is considered erroneous.}

\item{questionable_buffer}{A numeric value indicating a time buffer for
questionable values.}

\item{erroneous_buffer}{A numeric value indicating a time buffer for
erroneous values.}

\item{gap_range}{A numeric value indicating the range of hours between two
non reasonable values that will be coded as questionable or erroneous.
@param date_time A string indicating the column name of the date time vector.
@param value A string indicating the column name of the value vector.}
}
\value{
The original data frame sorted by date_time with a status_id column.
}
\description{
Time series data will be either classified as reasonable, questionable,
or erroneous in the status_id column.
}
\details{
The function only works on a single deployment of a logger. The
table output will be sorted by date_time.

The function will error if there are missing date_time values
missing.

The data is processed by:
\enumerate{
\item Classifying the temperature values based on their values
(questionable_min, questionable_max, erroneous_min, erroneous_max).
\item The
rate of change between adjacent values is calculate and values are
classified based on the rate parameters (questionable_rate,
erroneous_rate).
\item Adjacent values to questionable/erroneous are coded as
questionable/erroneous.
\item A buffer is applied that any value within the
buffer is classified as questionable/erroneous based on the buffer
parameters (questionable_buffer, erroneous_buffer).
\item Reasonable values
identified between two questionable/erroneous values are coded as
questionable/erroneous based on the gap hour difference allowed
(gap_range).
}
}
\examples{
data <- data.frame(
  date_time =
    as.POSIXct(c(
      "2021-05-07 08:00:00", "2021-05-07 09:00:00",
      "2021-05-07 10:00:00", "2021-05-07 11:00:00", "2021-05-07 12:00:00",
      "2021-05-07 13:00:00"
    )),
  water_temperature = c(4.124, 4.078, 4.102, 4.189, 4.243, 6.578)
)

classify_time_series_data(data, value = "water_temperature")
}
