% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{freshwaterfish}
\alias{freshwaterfish}
\title{BC Fish Data}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 161 rows and 17 columns.
}
\usage{
freshwaterfish
}
\description{
Curated data on the codes,
classification and conservation status of freshwater fishes in British Columbia.
}
\keyword{datasets}
