% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gdd-cf-data.R
\name{gdd_cf_data}
\alias{gdd_cf_data}
\title{Calculate Growing Degree Days (GDD) from a Data Frame}
\usage{
gdd_cf_data(
  x,
  start_date = as.Date("1972-01-01"),
  end_date = as.Date("1972-09-30"),
  ignore_truncation = FALSE,
  ...,
  span = 0,
  tails = FALSE
)
}
\arguments{
\item{x}{A data frame with two columns \code{date} and \code{temperature}.}

\item{start_date}{A Date scalar of the first date within each year to consider (the year
is ignored).}

\item{end_date}{A Date scalar of the end date for the growing degree days
(the year is ignored).
If \code{end_date} is less than \code{start_date} (ignoring the year) then the window is considered
to span two calendar years.}

\item{ignore_truncation}{A flag specifying whether to ignore start truncation (end truncation is always ignored).}

\item{...}{Additional arguments passed to \code{\link[=gsdd_cf]{gsdd_cf()}}.}

\item{span}{A whole number of the maximum span of missing values to interpolate.
If a gap exceeds the span none of the values in the gap are interpolated.}

\item{tails}{A flag specifying whether to fill in missing values at the
start and end by setting them to be the same value as the closest
adjacent non-missing value.}
}
\value{
A tibble with two columns \code{year} and \code{gdd}.
\code{year}, which is an integer vector, indicates the year in which the window
began and \code{gdd} which is a non-negative real number provides the GSDD
or a missing value if it cannot be calculated.
}
\description{
The GDD which is a the growing degree days to the end date
is calculated for each study year from a data frame.
with \code{date} and \code{temperature} columns.
\code{date}, which must be of class Date provides the dates and
\code{temperature} which must be a numeric vector provides the
mean daily water temperature in degrees centigrade.
}
\examples{
data <- teckfish::simulated_data
data$temperature <- data$synthetic
gdd_cf_data(data)
}
\seealso{
\code{\link[=gsdd_cf_data]{gsdd_cf_data()}}
}
