% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gsdd-cf.R
\name{gsdd_cf}
\alias{gsdd_cf}
\title{Calculate Growing Season Degree Days (GSDD)}
\usage{
gsdd_cf(
  x,
  na_trim = TRUE,
  ignore_truncation = FALSE,
  start_temp = 5,
  end_temp = 4,
  window_width = 7,
  pick = "all",
  quiet = FALSE
)
}
\arguments{
\item{x}{A numeric vector of the
mean daily water temperature values for the period
of interest in C. It must be consist of at least 90 and no more than
366 values.}

\item{na_trim}{A flag specifying whether to trim missing values
from the start and end of the time series. After trimming the time
series must consist of at least 90 values.}

\item{ignore_truncation}{A flag specifying whether to ignore truncation
of the mean daily water temperature vector
or a string of "start", "end", "none" or "both"
specifying which type of truncation to ignore.}

\item{start_temp}{A positive real number of the average water temperature
at the start of the growing season in C.}

\item{end_temp}{A positive real number of the average water temperature
at the end of the growing season in C. It must be greater than or equal to
the start temperature.}

\item{window_width}{A positive whole number of the
width of the rolling mean window in days. By default 7.}

\item{pick}{A string specifying whether to pick the
"longest", "shortest", "first" or "last" 'season' or the season with the
"biggest" or "smallest" GSDD. By default the returned value is the
sum of the GSDD values for "all" 'seasons'.}

\item{quiet}{A flag specifying whether to suppress warnings.}
}
\value{
A non-negative real number of the GSDD.
}
\description{
Growing Season Degree Days (GSDD) is a water temperature metric
that is a useful predictor of Cutthroat trout size at the
beginning of winter.
It is the accumulated thermal units (in C)
during the growing season based on the mean daily water temperature values.
}
\details{
By default the growing season is based on interpretation of
Coleman and Fausch (2007) who stated that

We defined the start of the growing season as the
beginning of the first week that average stream temperatures exceeded and
remained above 5C for the season;
the end of the growing season was defined as
the last day of the first week that
average stream temperature dropped below 4C.

For the purposes of the calculation week is assumed to refer to a seven day
rolling average as opposed to the calendar week,

If there are multiple growing 'seasons' within the same year then
the user has the option to pick the \code{"first"}/\code{"last"} or
\code{"longest"}/\code{"shortest"} season or the season with
the \code{"biggest"}/\code{"smallest"} GSDD.
By default the returned value is the sum of the GSDD values for \code{"all"} seasons.

If the user picks the \code{"longest"} season but there are multiple seasons
with the longest length then the candidate
season with the \code{"biggest"} GSDD is selected.
Conversely in the case of multiple \code{"shortest"} seasons then the
candidate with the \code{"smallest"} GSDD is selected.

If the user sets \code{na_trim = TRUE} then missing values (\code{NA}s) at the start and
end of the time series are trimmed prior to calculating GSDD.
If the time series still contains one or missing values then a missing value
is returned.
In this situation it is recommended that the user replace
the missing value(s) by interpolation or other methods rather than filtering
out the \code{NA}s.

Truncation occurs when the start and/or end
of the time series is part way through a growing season.
}
\examples{
gsdd_cf(c(rep(1, 10), rep(10, 20), rep(1, 200)))
gsdd_cf(teckfish::simulated_data$synthetic)
}
