% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gsdd-cf-data.R
\name{gsdd_cf_data}
\alias{gsdd_cf_data}
\title{Calculate Growing Season Degree Days (GSDD) from a Data Frame}
\usage{
gsdd_cf_data(x, start_month = 3, end_month = 10, ..., span = 3, tails = FALSE)
}
\arguments{
\item{x}{A data frame with two columns \code{date} and \code{water_temperature}.}

\item{start_month}{A whole number of the first month to include.}

\item{end_month}{A whole number of the last month to include.}

\item{...}{Additional arguments passed to \code{\link[=gsdd_cf]{gsdd_cf()}}.}

\item{span}{A whole number of the maximum span of missing values to interpolate.
If a gap exceeds the span none of the values are interpolate.}

\item{tails}{A flag specifying whether to fill in missing values at the
start and end by setting them to be the same value as the closest
adjacent non-missing value.}
}
\value{
A tibble with two columns \code{year} and \code{gsdd}.
\code{year}, which is an integer, indicates the year in which the window
began and \code{gsdd} which is a non-negative real number provides the GSDD
or a missing value if it cannot be calculated.
}
\description{
The GSDD is calculated for each study year from a data frame of
with a \code{date} and \code{water_temperature} values.
\code{date}, which must be of class \code{Date} provides the date and
\code{water_temperature} which must be a numeric vector provides the water temperature
in degrees centigrade. For additional information on
GSDD and the various arguments that can be passed via \code{...} see \code{\link[=gsdd_cf]{gsdd_cf()}}.
}
\examples{
x <- data.frame(water_temperature = teckfish::simulated_data$synthetic)
x$date <- dttr2::dtt_seq(as.Date("1970-01-01"), as.Date("1970-12-31"))
gsdd_cf_data(x)
gsdd_cf_data(x, ignore_truncation = TRUE)
gsdd_cf_data(x, end_month = 11)
}
