% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gsdd-vctr.R
\name{gsdd_vctr}
\alias{gsdd_vctr}
\title{Calculate Growing Season Degree Days (GSDD)}
\usage{
gsdd_vctr(x, ignore_truncation = FALSE, min_length = 274, msgs = TRUE)
}
\arguments{
\item{x}{A numeric vector of the
mean daily water temperature values for the period
of interest in C.}

\item{ignore_truncation}{A flag specifying whether to ignore truncation
of the mean daily water temperature vector
or a string of "start", "end", "none" (equivalent to FALSE) or "both"
(equivalent to TRUE) specifying which type of truncation to ignore.}

\item{min_length}{A whole number of the minimum length to consider.
Must be at least twice the window width but no more than 365.
If NULL then set to be the difference between \code{start_date} and \code{end_date}
(ignoring the year).}

\item{msgs}{A flag specifying whether to provide messages.}
}
\value{
A non-negative real number of the GSDD.
}
\description{
Calculate Growing Season Degree Days (GSDD)
}
\examples{
gsdd_vctr(c(rep(1, 10), rep(10, 20), rep(1, 200)))
gsdd_vctr(teckfish::simulated_data$synthetic)
}
\seealso{
\code{\link[gsdd:gsdd_vctr]{gsdd::gsdd_vctr()}}
}
