% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss-cf.R
\name{gss_cf}
\alias{gss_cf}
\title{Calculate Growing Seasons from a vector}
\usage{
gss_cf(
  x,
  ignore_truncation = FALSE,
  start_temp = 5,
  end_temp = 4,
  window_width = 7,
  msgs = TRUE
)
}
\arguments{
\item{x}{A numeric vector of the
mean daily water temperature values for the period
of interest in C. It must be consist of no more than
366 values.}

\item{ignore_truncation}{A flag specifying whether to ignore truncation
of the mean daily water temperature vector
or a string of "start", "end", "none" (equivalent to FALSE) or "both"
(equivalent to TRUE) specifying which type of truncation to ignore.}

\item{start_temp}{A positive real number of the average water temperature
at the start of the growing season in C.}

\item{end_temp}{A positive real number of the average water temperature
at the end of the growing season in C. It must be greater than or equal to
the start temperature.}

\item{window_width}{A positive whole number of the
width of the rolling mean window in days. By default 7.}

\item{msgs}{A flag specifying whether to provide messages.}
}
\value{
A tibble of the growing seasons with three columns.
\code{start_index} an integer vector of the start index,
\code{end_index} an integer vector of the end index and
\code{gsdd} an real vector of the growing season degree days.
}
\description{
Calculate Growing Seasons from a vector
}
\examples{
gss_cf(c(rep(1, 10), rep(10, 20), rep(1, 200)))
}
\seealso{
\code{\link[=gsdd_cf]{gsdd_cf()}}
}
