% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss-plot.R
\name{gss_plot}
\alias{gss_plot}
\title{Plot Growing Seasons (GSS)}
\usage{
gss_plot(
  x,
  min_length = 60,
  ignore_truncation = TRUE,
  pick = "all",
  latex = FALSE,
  nrow = NULL,
  ncol = NULL,
  msgs = TRUE
)
}
\arguments{
\item{x}{A data frame with two columns \code{date} and \code{temperature}.
\code{date}, which must be of class Date provides the dates and
\code{temperature} which must be a numeric vector provides the
mean daily water temperature in degrees centigrade.}

\item{min_length}{A whole number of the minimum length to consider.
Must be at least twice the window width but no more than 365.
If NULL then set to be the difference between \code{start_date} and \code{end_date}
(ignoring the year).}

\item{ignore_truncation}{A flag specifying whether to ignore truncation
of the mean daily water temperature vector
or a string of "start", "end", "none" (equivalent to FALSE) or "both"
(equivalent to TRUE) specifying which type of truncation to ignore.}

\item{pick}{A string specifying whether to pick the
"longest", "shortest", "first" or "last" 'season' or the season with the
"biggest" or "smallest" GSDD. By default the returned value is the
the GSDD value for the "longest" 'season'.}

\item{latex}{A flag specifying whether to use LaTeX to include degree
symbol in y-axis label.}

\item{nrow}{A count of the number of rows to facet by.}

\item{ncol}{A count of the number of columns to facet by.}

\item{msgs}{A flag specifying whether to provide messages.}
}
\description{
A wrapper on \code{\link[gsdd:gss_plot]{gsdd::gss_plot()}} to by default
plot all Growing Seasons ignoring truncation.
For more information see \code{\link[gsdd:gss_plot]{gsdd::gss_plot()}}.
}
\examples{
gss_plot(gsdd::temperature_data)
}
\seealso{
\code{\link[gsdd:gss_plot]{gsdd::gss_plot()}} and \code{\link[=gss]{gss()}}.
}
