% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/params.R
\name{params}
\alias{params}
\alias{parameters}
\alias{arguments}
\alias{args}
\title{Parameter Descriptions}
\arguments{
\item{...}{These dots are for future extensions and must be empty.}

\item{data}{A data frame.}

\item{date_time}{A string indicating the column name of the POSIXct vector.}

\item{gap_range}{A numeric value indicating the number of hours between two
non reasonable values that will be coded as questionable or erroneous.}

\item{erroneous_buffer}{A numeric value indicating the number of hours buffer for
erroneous values.}

\item{erroneous_max}{A numeric value indicating the upper bound of the
erroneous range of temperature values.}

\item{erroneous_min}{A numeric value indicating the lower bound of the
erroneous range of temperature values.}

\item{erroneous_rate}{A numeric value indicating the rate of change
(temperature per hour) of temperature values that is considered erroneous.}

\item{questionable_buffer}{A numeric value indicating the buffer in hours for
questionable values.}

\item{questionable_min}{A numeric value indicating the lower bound of the
questionable range of temperature values.}

\item{questionable_max}{A numeric value indicating the upper bound of the
questionable range of temperature values.}

\item{questionable_rate}{A numeric value indicating the rate of change
(temperature per hour) of temperature values that is considered
questionable.}

\item{value}{A string indicating the column name of the value vector.}
}
\description{
Default parameter descriptions which may be overridden in individual
functions.
}
\details{
A flag is a non-missing logical scalar.

A string is a non-missing character scalar.
}
\keyword{internal}
