% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{simulated_data}
\alias{simulated_data}
\title{Sample Data for Population A}
\format{
A tibble with columns:
\describe{
\item{Date}{date for each calendar year}
\item{synthetic}{synthetic yearly temperature data for demonstration and testing}
}
}
\usage{
simulated_data
}
\description{
The data contains recruitment information for boreal caribou population B
which spans 26 years.
}
\details{
The data is released under the \href{https://open.alberta.ca/licence}{Open Government Licence - Alberta}
}
\keyword{datasets}
