% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/teckfish-package.R
\docType{package}
\name{teckfish-package}
\alias{teckfish}
\alias{teckfish-package}
\title{teckfish: R Tools for Teck Fish and Aquatic Sciences Projects}
\description{
Provides functions linearly interpolate missing values and calculate growing season degree days and growing degree days from daily water temperature data. It also exports fish codes for British Columbia and Alberta from the `fishbc` package.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://poissonconsulting.github.io/teckfish/}
  \item \url{https://github.com/poissonconsulting/teckfish}
  \item Report bugs at \url{https://github.com/poissonconsulting/teckfish/issues/}
}

}
\author{
\strong{Maintainer}: Joe Thorley \email{joe@poissonconsulting.ca} (\href{https://orcid.org/0000-0002-7683-4592}{ORCID})

Authors:
\itemize{
  \item Ayla Pearson \email{ayla@poissonconsulting.ca} (\href{https://orcid.org/0000-0001-7388-1222}{ORCID})
}

Other contributors:
\itemize{
  \item Sarah Lyons (\href{https://orcid.org/0000-0002-6745-6796}{ORCID}) [contributor]
  \item Bronwen Lewis [contributor]
  \item Jill Brooks [contributor]
  \item Andrew Harwood [contributor]
  \item Sebastian Dalgarno (\href{https://orcid.org/0000-0000-0000-0000}{ORCID}) [contributor]
  \item Teck Resources [funder, copyright holder]
}

}
\keyword{internal}
