% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_gsdd.R
\name{calculate_gsdd}
\alias{calculate_gsdd}
\title{Calculate Growing Season Degree Days (GSDD)}
\usage{
calculate_gsdd(
  x,
  window_width = 7,
  start_temp = 5,
  end_temp = 4,
  n_consecutive = 5
)
}
\arguments{
\item{x}{A vector of numeric temperature data.}

\item{window_width}{A positive whole number indicating the width of rolling mean window.}

\item{start_temp}{A number indicating the threshold rolling average temperature to define the start of the growing season.}

\item{end_temp}{A number indicating the weekly temperature rolling average threshold to define the end of growing season.}

\item{n_consecutive}{A positive whole number indicating the number of consecutive rolling average days above and below cut off temperatures to begin GSDD calculations.}
}
\value{
A number of the cumulative GSDD.
}
\description{
GSDD is calculated by summing daily temperatures during the growing season.
The start of the growing season is defined by the rolling mean temperature remaining above the start_temp for n_consecutive days.
The end of the growing season is defined by the rolling mean temperature remaining below the end_temp for n_consecutive days.
}
\examples{
set.seed(13)
day <- 1:365
temperature <- -15 * cos((2*pi / 365) * (day-10)) + rnorm(365, mean = 10, sd = .5)
calculate_gsdd(x = temperature, window_width = 7, start_temp = 5, end_temp = 4, n_consecutive = 5)

}
