% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/teckfish-package.R
\docType{package}
\name{teckfish-package}
\alias{teckfish}
\alias{teckfish-package}
\title{teckfish: Calculate Common Growth Metrics}
\description{
Allows reproducible calculation of fish growth metrics such as seasonal growing degree days.
}
\author{
\strong{Maintainer}: Joe Thorley \email{joe@poissonconsulting.ca} (\href{https://orcid.org/0000-0002-7683-4592}{ORCID})

Authors:
\itemize{
  \item Sarah Lyons (\href{https://orcid.org/0000-0002-6745-6796}{ORCID})
}

Other contributors:
\itemize{
  \item Sebastian Dalgarno (\href{https://orcid.org/0000-0000-0000-0000}{ORCID}) [contributor]
}

}
\keyword{internal}
