% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/teckfish-package.R
\docType{package}
\name{teckfish-package}
\alias{teckfish}
\alias{teckfish-package}
\title{teckfish: R Tools for Teck Fish and Aquatic Sciences Projects}
\description{
Currently only includes function to calculate growing season degree days from daily water temperature data.
}
\author{
\strong{Maintainer}: Joe Thorley \email{joe@poissonconsulting.ca} (\href{https://orcid.org/0000-0002-7683-4592}{ORCID})

Authors:
\itemize{
  \item Sarah Lyons (\href{https://orcid.org/0000-0002-6745-6796}{ORCID})
  \item Ayla Pearson \email{ayla@poissonconsulting.ca} (\href{https://orcid.org/0000-0001-7388-1222}{ORCID})
}

Other contributors:
\itemize{
  \item Sebastian Dalgarno (\href{https://orcid.org/0000-0000-0000-0000}{ORCID}) [contributor]
  \item Teck Resources [funder, copyright holder]
}

}
\keyword{internal}
